MLG - 8:10 PM - Changes error for task 4

+++ 0x03-asynchronous/4-get_sum.dart
@@ -28,6 +28,6 @@ Future<double> calculateTotal() async {
     // Return total of items
     return total;
   } catch (error) {
-    return -1.0;
+    return -1.toDouble();
   }
 }
